fix all values cache for companies
